Remove files that are no longer needed
